+++ assignment-2/src/main/java/com/lucentblock/assignment2/security/authentication/AuthenticationService.java
@@ -196,28 +196,6 @@ public class AuthenticationService {
         throw new AlreadyVerifiedUserException(retrievedUser.getEmail());
     }
 
-    public UserInfoDTO fetchUserInfo(String userEmail) {
-        User user = userRepository.findByEmailAndDeletedAtIsNull(userEmail).orElseThrow(() -> new UsernameNotFoundException(userEmail));
-
-        return UserInfoDTO.UserEntityToUserInfoDTO(user);
-    }
-
-    public UserInfoDTO updateUserInfo(UserInfoDTO userInfoDTO) {
-        User user = userRepository.findByEmailAndDeletedAtIsNull(userInfoDTO.getUserEmail())
-                .orElseThrow(() -> new UsernameNotFoundException(userInfoDTO.getUserEmail()));
-
-        User savedUser = userRepository.saveAndFlush(user.UpdateUserBasedOnUserInfoDTO(userInfoDTO));
-        return UserInfoDTO.UserEntityToUserInfoDTO(savedUser);
-    }
-
-    public ResponseEntity deleteUser(String userEmail) {
-        User retrievedUser = userRepository.findByEmailAndDeletedAtIsNull(userEmail).orElseThrow(() -> new UsernameNotFoundException(userEmail));
-        retrievedUser.setDeletedAt(LocalDateTime.now());
-        userRepository.saveAndFlush(retrievedUser);
-
-        return ResponseEntity.ok().build();
-    }
-
     private void makeLoginChallenge(User user, boolean isSuccessful) {
         loginChallengeRepository.save(LoginChallenge.builder()
                 .user(user)
