--- assignment-2/src/test/java/com/lucentblock/assignment2/security/authentication/DeleteUserInfoTest.java
@@ -1,75 +1,0 @@
-package com.lucentblock.assignment2.security.authentication;
-
-import com.lucentblock.assignment2.entity.Role;
-import com.lucentblock.assignment2.entity.User;
-import com.lucentblock.assignment2.repository.LoginChallengeRepository;
-import com.lucentblock.assignment2.repository.SignupCodeChallengeRepository;
-import com.lucentblock.assignment2.repository.UserRepository;
-import com.lucentblock.assignment2.security.authentication.jwt.JwtRefreshService;
-import com.lucentblock.assignment2.security.authentication.jwt.JwtService;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.http.ResponseEntity;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-import java.time.LocalDateTime;
-import java.util.Optional;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.BDDMockito.given;
-
-@ExtendWith(MockitoExtension.class)
-public class DeleteUserInfoTest {
-    @Mock
-    private UserRepository userRepository;
-
-    @InjectMocks
-    private AuthenticationService authService;
-
-    private User user;
-
-    @BeforeEach
-    void setup() {
-        user = User.builder()
-                .name("testName")
-                .email("test@test.com")
-                .password("testPassword")
-                .role(Role.ROLE_USER)
-                .isEmailVerified(false)
-                .phoneNumber("testPhoneNumber")
-                .refreshToken("refresh_token")
-                .createdAt(LocalDateTime.now())
-                .build();
-    }
-
-    @Test
-    @DisplayName("존재하는 회원에 대해 삭제를 수행할 수 있다.")
-    void deleteUser() {
-        // given
-        given(userRepository.findByEmailAndDeletedAtIsNull(user.getEmail())).willReturn(Optional.of(user));
-
-        // when
-        ResponseEntity response = authService.deleteUser(user.getEmail());
-
-        // then
-        assertEquals(true, response.getStatusCode().is2xxSuccessful());
-    }
-
-    @Test
-    @DisplayName("존재하지 않는 회원에 대해 삭제를 요청할 경우 UsernameNotFound Exception 이 발생한다.")
-    void deleteUserDoesNotExist() {
-        // given
-        given(userRepository.findByEmailAndDeletedAtIsNull(user.getEmail())).willThrow(new UsernameNotFoundException(user.getEmail()));
-
-        // when & then
-        assertThrows(UsernameNotFoundException.class, () -> authService.deleteUser(user.getEmail()));
-    }
-}
