+++ assignment-2/src/test/java/com/lucentblock/assignment2/security/authentication/service/RegisterTest.java
@@ -1,10 +1,11 @@
-package com.lucentblock.assignment2.security.authentication;
+package com.lucentblock.assignment2.security.authentication.service;
 
 import com.lucentblock.assignment2.entity.Role;
 import com.lucentblock.assignment2.entity.User;
 import com.lucentblock.assignment2.repository.LoginChallengeRepository;
 import com.lucentblock.assignment2.repository.SignupCodeChallengeRepository;
 import com.lucentblock.assignment2.repository.UserRepository;
+import com.lucentblock.assignment2.security.authentication.AuthenticationService;
 import com.lucentblock.assignment2.security.authentication.jwt.JwtRefreshService;
 import com.lucentblock.assignment2.security.authentication.jwt.JwtService;
 import com.lucentblock.assignment2.security.exception.UserDuplicateException;
